DataSeeder

+++ CalorieTracker.Domains/Product.cs
@@ -2,12 +2,12 @@
 {
     public class Product : BaseNamedEntity
     {
-        public int CaloriePer100g { get; set; }
+        public double CaloriePer100g { get; set; }
 
-        public int ProteinPer100g { get; set; }
+        public double ProteinPer100g { get; set; }
 
-        public int FatPer100g { get; set; }
+        public double FatPer100g { get; set; }
 
-        public int CarbohydratePer100g { get; set; }
+        public double CarbohydratePer100g { get; set; }
     }
 }
